Props are added in the center of the camera in editor. Enemies now can be selectable in editor. Created folders in xml and textures for enemies, animatedProps, characters, projectiles. Projectiles are now animated entities. Added the visual and basic behavior to grape.xml Added basic behavior to the player. Now the scale of entities loaded from texture or xml is automatically assigned.

+++ MyGame/MyGame/code/Render & Effects/AnimatedTexture.cs
@@ -14,8 +14,8 @@ namespace MyGame
         public Texture2D texture { get; set; }
         public int columns { get; set; }
         public int rows { get; set; }
-        public int frameWidth { get; set; }
-        public int frameHeight { get; set; }
+        public float frameWidth { get; set; }
+        public float frameHeight { get; set; }
         public float frameWidthUV { get; set; }
         public float frameHeightUV { get; set; }
     }
